change variable type

+++ src/main/java/functions/Calculator.java
@@ -101,8 +101,8 @@ public class Calculator<T> {
         System.out.println(x+""+answer+""+answer1+""+answe4);
     }
     public Integer Layer1(Integer i){
-        int ii = B.i;
-        int iii = functions.special.B.i;
+        long ii = B.i;
+        long iii = functions.special.B.i;
         Integer j = B.Layer2_RENAME(i);
         return j;
     }
